t08_flask_mysql/app/requirements.txt
flask
waitress
flask-sqlalchemy
mysqlclient
sqlalchemy-utils
jinja2
jsonify
flask-rests